https://jira.ccpgames.com/browse/GFX-4659
Fix compilation problems.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1396772]

+++ Tr2SwapChain.cpp
@@ -8,7 +8,7 @@ Tr2SwapChain::Tr2SwapChain( IRoot* lockobj )
 {
 }
 
-bool Tr2SwapChain::CreateForWindow( int windowHandle )
+bool Tr2SwapChain::CreateForWindow( size_t windowHandle )
 {
 	m_windowHandle = (Tr2WindowHandle)windowHandle;
 	return PrepareResources();
